refactor: remove unused components & files
