--- src/component/Button/index.ts
@@ -1,2 +1,0 @@
-export { Button } from "./Button";
-export { NavLink } from "./NavLink";
