--- src/layout/Footer.tsx
@@ -1,10 +1,0 @@
-import type { VFC } from "react";
-
-/**
- * @package
- */
-export const Footer: VFC = () => (
-  <div>
-    <small>&copy; 20xx example</small>
-  </div>
-);
