--- src/pages/about/index.tsx
@@ -1,3 +1,0 @@
-import type { VFC } from "react";
-
-export const About: VFC = () => <h2>About!</h2>;
